# openapi-atspaces.yaml
openapi: 3.0.3
info:
  title: At Spaces API
  version: 1.0.0
  description: |
    API contract for At Spaces (NestJS + Prisma + PostgreSQL + Redis).
    Roles: customer, vendor, admin. Access token in JSON, refresh token in HttpOnly cookie.
servers:
  - url: http://localhost:3000/api
  - url: https://<domain>/api
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UserPublic:
      type: object
      properties:
        id: { type: integer, example: 123 }
        role: { type: string, enum: [customer, vendor, admin] }
        fullName: { type: string, example: "John Doe" }
    AuthLoginResponse:
      type: object
      properties:
        accessToken: { type: string, example: "jwt" }
        user:
          $ref: '#/components/schemas/UserPublic'
    RegisterEmailRequest:
      type: object
      required: [fullName, email, password]
      properties:
        fullName: { type: string }
        email: { type: string, format: email }
        password: { type: string, format: password }
    RegisterPhoneRequest:
      type: object
      required: [fullName, phoneNumber]
      properties:
        fullName: { type: string }
        phoneNumber: { type: string, example: "+9627xxxxxxxx" }
    VerifyOtpRequest:
      type: object
      required: [phoneNumber, otpCode, purpose]
      properties:
        phoneNumber: { type: string, example: "+9627xxxxxxxx" }
        otpCode: { type: string, example: "123456" }
        purpose: { type: string, enum: [login, signup, verify] }
    ResendOtpRequest:
      type: object
      required: [phoneNumber, purpose]
      properties:
        phoneNumber: { type: string, example: "+9627xxxxxxxx" }
        purpose: { type: string, enum: [login, signup, verify] }
    LoginEmailRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, format: password }
    ForgotPasswordRequest:
      type: object
      required: [email]
      properties:
        email: { type: string, format: email }
    ResetPasswordRequest:
      type: object
      required: [resetToken, newPassword]
      properties:
        resetToken: { type: string }
        newPassword: { type: string, format: password }
    AdminResetPasswordRequest:
      type: object
      required: [resetToken, newPassword, totpCode]
      properties:
        resetToken: { type: string }
        newPassword: { type: string, format: password }
        totpCode: { type: string, example: "123456" }
    AdminLoginRequest:
      type: object
      required: [email, password, captchaToken]
      properties:
        email: { type: string, format: email }
        password: { type: string, format: password }
        captchaToken: { type: string }
    AdminLoginResponse:
      type: object
      properties:
        preAuthToken: { type: string }
        mfaRequired: { type: boolean, example: true }
...
paths:
...
  /auth/customer/register-email:
    post:
      tags: [Auth - Customer]
      summary: Register customer with email
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RegisterEmailRequest' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  userId: { type: integer }
                  message: { type: string }
  /auth/customer/register-phone:
    post:
      tags: [Auth - Customer]
      summary: Register customer with phone (OTP)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RegisterPhoneRequest' }
      responses:
        '202':
          description: OTP sent
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
  /auth/customer/verify-otp:
    post:
      tags: [Auth - Customer]
      summary: Verify OTP for customer
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/VerifyOtpRequest' }
      responses:
        '200':
          description: Authenticated
          headers:
            Set-Cookie:
              description: Refresh cookie (HttpOnly)
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthLoginResponse' }
  /auth/customer/login-email:
    post:
      tags: [Auth - Customer]
      summary: Customer login (email/password)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginEmailRequest' }
      responses:
        '200':
          description: Authenticated
          headers:
            Set-Cookie:
              description: Refresh cookie (HttpOnly)
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthLoginResponse' }
  /auth/customer/resend-otp:
    post:
      tags: [Auth - Customer]
      summary: Resend OTP to phone
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ResendOtpRequest' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
...
  /auth/vendor/login:
    post:
      tags: [Auth - Vendor]
      summary: Vendor login (email/password)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginEmailRequest' }
      responses:
        '200':
          description: Authenticated
          headers:
            Set-Cookie:
              description: Refresh cookie (HttpOnly)
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthLoginResponse' }
...
  /vendors/register:
    post:
      tags: [Vendors]
      summary: Vendor registration (requires admin approval)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [fullName, email, password, branch]
              properties:
                fullName: { type: string }
                email: { type: string, format: email }
                password: { type: string, format: password }
                branch:
                  type: object
                  required: [name, city, address]
                  properties:
                    name: { type: string }
                    city: { type: string }
                    address: { type: string }
                    latitude: { type: number, format: float, nullable: true }
                    longitude: { type: number, format: float, nullable: true }
      responses:
        '202':
          description: Submitted for approval
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
  /admin/auth/login:
    post:
      tags: [Auth - Admin]
      summary: Admin login step 1 (password + captcha)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AdminLoginRequest' }
      responses:
        '200':
          description: Pre-auth token
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminLoginResponse' }
...
  /auth/customer/forgot-password:
    post:
      tags: [Auth - Customer]
      summary: Customer forgot password
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ForgotPasswordRequest' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
  /auth/customer/reset-password:
    post:
      tags: [Auth - Customer]
      summary: Customer reset password
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ResetPasswordRequest' }
      responses:
        '200':
          description: Password updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
  /auth/vendor/forgot-password:
    post:
      tags: [Auth - Vendor]
      summary: Vendor forgot password
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ForgotPasswordRequest' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
  /auth/vendor/reset-password:
    post:
      tags: [Auth - Vendor]
      summary: Vendor reset password
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ResetPasswordRequest' }
      responses:
        '200':
          description: Password updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
  /admin/auth/forgot-password:
    post:
      tags: [Auth - Admin]
      summary: Admin forgot password
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ForgotPasswordRequest' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
  /admin/auth/reset-password:
    post:
      tags: [Auth - Admin]
      summary: Admin reset password (requires TOTP)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AdminResetPasswordRequest' }
      responses:
        '200':
          description: Password updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
...